Change unordered marker from asterik to dash
The previously used asterisk `*` have been changed to dashes `-`. The
main reason is that asterisks `*` can be confused for bold/italic
markers. There was also a very long discussion (1) for Prettier (2) to
also change the default to dashes to same reasons which was then
adapted (3).
There are some other reasons like the alignment with the YAML spec which
uses dashes for list items.

This guide has now also adapted to this improvement by changing the
default unordered list marker (4) from asterisks `*` to dashes `-`.

References:

  (1) https://github.com/prettier/prettier/issues/4251#issue-310943939
  (2) https://prettier.io
  (3) https://github.com/prettier/prettier/pull/4440
  (4) https://arcticicestudio.github.io/styleguide-markdown/rules/lists.html

GH-3