`svengreb` GitHub account and `@svengreb` npm package scope migration (#64)
With the retirement of the "Arctic Ice Studio" personal & Nord project
brand [1] this project also moved to the "real-in-person" identity
"Sven Greb" both in the context of the repository to the `svengreb`
GitHub account [2] and the `@svengreb` npm package scope.
During this migration the currently only npm package
`@arcticicestudio/remark-preset-lint` [3] has been deprecated in favor
of the new and upcoming `@svengreb/remark-preset-lint` [3] package that
has been published afterwards.

Also the visual representation of this style guide that was published
before [6] through the way too outdated and deprecated GitBook [5] major
version `2` has now been unpublished and removed. The documentations and
references are updated to use the GitHub repository with the Markdown
rendering instead for now until a custom website has been implemented
using a modern "TechStack" like [Next.js][7]. A simple HTTP redirection
page has been published through the `gh-pages` Git branch that redirects
to `https://github.com/svengreb/styleguide-markdown` to catch possible
users and make them aware of the migration.

[1]: https://github.com/orgs/nordtheme/discussions/183#retire-arctic-ice-studio-as-nord-brand
[2]: https://github.com/svengreb
[3]: https://www.npmjs.com/package/@arcticicestudio/remark-preset-lint
[4]: https://www.npmjs.com/package/@svengreb/remark-preset-lint
[5]: https://www.gitbook.com
[6]: https://arcticicestudio.github.io/styleguide-markdown
[7]: https://nextjs.org

GH-63